gpt-chatbot/requirements.txt
streamlit
streamlit-chat
openai
python-dotenv
